Add UnitReadoutDialog based on AbstractDialog
